+++ changelog.md
@@ -3,4 +3,5 @@
 * Leaf litters now generate under every naturally generated white oak trees, even in plains.
 * Wandering Trader now sells 8 White Oak Logs as a special offer.
 * Rainbow Rose now spread using bone meal.
+* Iron blocks now makes iron sounds.
 * Changed Rainbow Petals ID from `rainbow_petal` to `rainbow_petals`